WIP on refactor: common functionality for execution moved from resttest class
