+++ pyresttest/test_parsing.py
@@ -127,13 +127,41 @@ class TestParsing(unittest.TestCase):
 
         self.assertEqual({'newval': 'cherries'}, safe_to_json(Special()))
 
-    def test_run_configure(self):
-        """ Test the configure function use """
-        converter = safe_to_bool
-        pass
+    def test_cmdline_args_parsing_basic(self):
+        cmdline = [
+            'my_url', 'my_test_filename',
+            '--print-bodies', 'True'
+        ]
+        args = parse_command_line_args(cmdline)
+        self.assertEqual('my_url', args['url'])
+        self.assertEqual('my_test_filename', args['test'])
+        self.assertEqual('True', args['print_bodies'])
 
-    def test_configure(self):
-        """ Do stuff here """
-        pass
+    def test_cmdline_args_parsing_positional(self):
+        """ Tests cases where test and url are from named arguments, not positional """
+
+        cmdline = [
+            '--url', 'my_url',
+            '--test', 'my_test_filename',
+        ]
+
+        args = parse_command_line_args(cmdline)
+        self.assertEqual('my_url', args['url'])
+        self.assertEqual('my_test_filename', args['test'])
+
+        # url from position arg, test as named arg
+        del cmdline[0]
+        args = parse_command_line_args(cmdline)
+        self.assertEqual('my_url', args['url'])
+        self.assertEqual('my_test_filename', args['test'])
+
+        cmdline = [
+            '--url', 'my_url',
+            'my_test_filename',
+        ]
+        args = parse_command_line_args(cmdline)
+        self.assertEqual('my_url', args['url'])
+        self.assertEqual('my_test_filename', args['test'])
+
 if __name__ == '__main__':
     unittest.main()
