+++ setup.py
@@ -36,7 +36,7 @@ setup(name='pyresttest',
       ],
       py_modules=['pyresttest.resttest', 'pyresttest.generators', 'pyresttest.binding',
                   'pyresttest.parsing', 'pyresttest.validators', 'pyresttest.contenthandling',
-                  'pyresttest.benchmarks', 'pyresttest.tests', 
+                  'pyresttest.benchmarks', 'pyresttest.tests', 'pyresttest.macros', 
                   'pyresttest.six',
                   'pyresttest.ext.validator_jsonschema',
                   'pyresttest.ext.extractor_jmespath'],
